superrag/pyproject.toml
[tool.poetry]
name = "superrag"
version = "0.1.0"
description = "Instant RAG data API"
authors = ["Simrat Hanspal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
cython = "0.29.36"
gql = "3.4.0"
openai = "^1.3.0"
tenacity = "^8.2.3"
requests = "^2.31.0"
requests-toolbelt = "^1.0.0"
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
setuptools = "^68.2.2"
langchain = "^0.0.339"
langchain-experimental = "^0.0.42"
llama-index = "^0.9.5"
sqlalchemy = "^2.0.23"
psycopg2 = "2.9.6"
pandas = "^2.1.3"
scikit-learn = "^1.3.2"
argparse = "^1.4.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
